workers/dlq_replayer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
structlog==23.2.0
redis==5.0.1